phy: hdmigtctrl: Update the GT values
The GT values need to be programmed based on the Tx PLL and Rx PLL type.
These can either be LCPLL or RPLL type. The Tx PLL type and Rx PLL type
will not be same. This patch updates the values to be programmed in GT
for Versal.

The Reset PLL register reset in HDMI GT Controller (0x14 offset) is
removed too.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>